Improve Grids model and implement repositories for it.

+++ Backend/GridsApi/Service/Repositories/Grids.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.Data.SqlClient;
+
+using System.Diagnostics;
+
 using Service.DatabaseModels;
-using System.Runtime.CompilerServices;
 
 namespace Service.Repositories
 {
@@ -13,19 +15,32 @@ namespace Service.Repositories
             _context = context;
         }
 
-        public async Task AddAsync(Grid entity)
+        public async Task<int> AddAsync(Grid entity)
         {
+            int? gridId = null;
+
             try
             {
-                _context.Grids.Add(entity);
-                await _context.SaveChangesAsync();
+                var existingGrid = _context.Grids.Where(g => g.Name == entity.Name).FirstOrDefault();
+
+                if (existingGrid != null)
+                {
+                    gridId = existingGrid.Id;
+                }
+                else
+                {
+                    _context.Grids.Add(entity);
+                    await _context.SaveChangesAsync();
+
+                    gridId = entity.Id;
+                }
             }
             catch (SqlException ex)
             {
                 const int violationOfUniqueConstraint = 2601;
                 if (ex.Number == violationOfUniqueConstraint)
                 {
-                    // Grid already exists, proceed forward.
+                    throw new Exception($"Failed to add new grid to database. Probably due to race between 2 hosts trying to add grid at the same time. Please, retry... Details: {ex.Message}");
                 }
                 else
                 {
@@ -36,14 +51,24 @@ namespace Service.Repositories
             {
                 throw new Exception($"Failed to add new grid to database with error: {ex.Message}");
             }
+
+            Debug.Assert(gridId != null);
+            return (int)gridId;
         }
 
-        public async Task Remove(Grid entity)
+        public async Task RemoveAsync(Grid entity)
         {
             try
             {
-                _context.Grids.Remove(entity);
-                await _context.SaveChangesAsync();
+                if (!_context.Hosts.Any(h => h.GridId == entity.Id))
+                {
+                    _context.Grids.Remove(entity);
+                    await _context.SaveChangesAsync();
+                }
+                else
+                {
+                    throw new Exception($"There are hosts running in {entity.Name} grid. Will not remove.");
+                }
             }
             catch (Exception ex)
             {
@@ -51,22 +76,22 @@ namespace Service.Repositories
             }
         }
 
-        public Grid GetById(int id)
+        public Task<Grid> GetByIdAsync(int id)
         {
             throw new NotImplementedException();
         }
 
-        public Grid GetByName(string name)
+        public Task<Grid> GetByNameAsync(string name)
         {
             throw new NotImplementedException();
         }
 
-        public IEnumerable<Grid> List()
+        public Task<IEnumerable<Grid>> ListAsync()
         {
             throw new NotImplementedException();
         }
 
-        public IEnumerable<Grid> List(System.Linq.Expressions.Expression<Func<Grid, bool>> predicate)
+        public Task<IEnumerable<Grid>> ListAsync(System.Linq.Expressions.Expression<Func<Grid, bool>> predicate)
         {
             throw new NotImplementedException();
         }
